Working on testing login and register in AuthController
Changed player id from string to int, so db needs to be deleted and created again.
Alot of null warning that might need to be supressed.
May be some methods that are broken because of player id being a int now.

IMPORTANT: When parsing claims in token, you need to parse the string to a int